Revert "128 in separate storage does not give a lot of benefit"
This reverts commit 3cd27bb91363d4aaa05cf66c23ab35bbaa951cbb.

+++ silo-core/contracts/interfaces/ISilo.sol
@@ -47,7 +47,7 @@ interface ISilo is IERC4626, IERC3156FlashLender, ISiloLiquidation {
     /// It also includes token amount that has been borrowed.
     /// - DEBT: Amount of asset token that has been borrowed plus accrued interest.
     struct Assets {
-        uint128 assets;
+        uint256 assets;
     }
 
     /// @dev Storage struct that holds all required data for a single token market
